関数に切り分けた
